BACKLOG-20851: update CE url, share expandTree logic
